Added new classes  ChppGikSolverBasic and ChppGikSolverHumanoid.
ChppGikSolverBasic is basically a solver of prioritized linear systems. It does not need a robot object.
ChppGikSolverHumanoid is tailored to humanoid robot inverse kinematics.

ChppGikSolver is maintained in the package for now but SHOULD DISAPPEAR in the future. Please use the ChppGikSolverBasic or ChppGikSolverHumanoid classes instead.

Added algorithm to compute convex hull of footprints on ground. This is used in the coming package: hppLocalStepper.