Version 2 - Codename "GIK N' Furious"
The package is upgraded to version 2 since it presents a few incompaibilities with earlier versions. The good news is that computation times are halfed compared to last version.

Among the changes:

The priority solver is unified to ChppGikSolver (previously ChppGikSolverRobotAttached and ChppGikSolverRobotFree, these no longer exist).

There is no need to record a "FixedJoint" in the robot any more, nor to compute the fixedfoot jacobian prior to entering the constraints in the gik solver.
One, however, need to make sure the constraints have the same jacobian root set(see new version of abstractGikTask package). A method prepare() is added to ChppGikSolver class to take care of jacobian root setting and value and jacobian computations in the stack of constraints prior to solving.
The root for the the jacobian computation can be set in ChppGikSolver. The weights that are entered in ChppGikSolver still determine which degrees of freedom are put to contribution. The first 6 degrees of freedom in that weight vector denote the degrees of freedom of the body attached to the joint selected as root of jacobians. The rest correspond to the degrees of freedom holding the same rank in the robot's configuration.

Phew!

Now, the mask vectors returned by objects of class ChppGikmaskFactory have now the same size as the robot's configuration.

In ChppGikConfigurationConstraint's construction arguments, the mask has now the same as the target, that is to say the size of the robot's config.

There is still room for improvement: in the current state, the robot's root is supposed to be a freeflyer whose 6 degrees of freedom are placed at the head of the configuration vector and necessarily representing [x y z roll pitch yaw].

Version passed a reasanoble number of tests but not sure if it's stable enough for manipulation.

Enjoy.